add  components folder
